refactor(imports): update imports after domain restructuring
- fix paths after moving files
- ensure project builds correctly
