+++ src/types/drawerParamTypes.ts
@@ -1,0 +1,6 @@
+export type DrawerParamList = {
+  HomeScreen: undefined;
+  Account: undefined;
+  Activity: undefined;
+  Cars: undefined;
+};
